80% done

+++ database/migrations/2024_10_10_064245_create_pakans_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('pakans', function (Blueprint $table) {
             $table->id();
             $table->string('nama_pakan');
-            $table->unsignedBigInteger('qty');
-            $table->unsignedBigInteger('harga');
+            $table->decimal('qty', 15, 3);
+            $table->decimal('harga', 15, 3);
             $table->timestamps();
         });
     }
